Add Inventory Creation Services and Refactor Component
Introduce InventoryCreateProcessorService and InventoryCreateFacade to handle business logic and state management for the inventory creation process. Simplify InventoryCreateComponent by delegating responsibilities to the newly created services, enhancing code readability and maintainability.
